requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai>=1.54.0
httpx>=0.27.0,<0.28.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pillow==10.4.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
requests==2.32.3
alembic==1.13.3
